refactor: Entity 삭제
