+++ src/main/java/chess/domain/game/ChessGame.java
@@ -5,7 +5,6 @@ import chess.domain.piece.Color;
 import chess.domain.piece.Piece;
 import chess.domain.piece.Pieces;
 import chess.domain.position.Position;
-import chess.entity.Room;
 import java.util.List;
 
 public class ChessGame {
